+++ packages/cli/src/merge-map.ts
@@ -397,6 +397,7 @@ export const VARIANTS: ReadonlySet<string> = new Set([
 	"@pc:",
 	"@prm:",
 	"@sm:",
+	"@st:",
 	"@xl:",
 	"@xs:",
 	"@xxl:",